Adding the caps

+++ repos/fdbt-site/src/pages/api/deleteCap.ts
@@ -1,7 +1,11 @@
 import { NextApiResponse } from 'next';
-import { deleteCap } from '../../data/auroradb';
+import { deleteCap, getAllProductsByNoc, getCapByNocAndId } from '../../data/auroradb';
 import { redirectToError, redirectTo, getAndValidateNoc } from '../../utils/apiUtils/index';
-import { NextApiRequestWithSession } from '../../interfaces';
+import { CapInfo, ErrorInfo, NextApiRequestWithSession } from '../../interfaces';
+import { getProductsMatchingJson } from '../../data/s3';
+import { updateSessionAttribute } from '../../utils/sessions';
+import { VIEW_CAP } from '../../constants/attributes';
+import { TicketWithIds } from 'src/interfaces/matchingJsonTypes';
 
 export default async (req: NextApiRequestWithSession, res: NextApiResponse): Promise<void> => {
     try {
@@ -14,6 +18,26 @@ export default async (req: NextApiRequestWithSession, res: NextApiResponse): Pro
         }
 
         const noc = getAndValidateNoc(req, res);
+
+        const products = await getAllProductsByNoc(noc);
+        const matchingJsonLinks = products.map((product) => product.matchingJsonLink);
+        const tickets: TicketWithIds[] = await Promise.all(
+            matchingJsonLinks.map(async (link) => {
+                return await getProductsMatchingJson(link);
+            }),
+        );
+
+        const productsUsingCap = tickets.filter((t) => t.cap && t.cap.id === id);
+
+        if (productsUsingCap.length > 0) {
+            const capInfo = (await getCapByNocAndId(noc, id)) as CapInfo;
+            const { name } = capInfo.cap;
+            const errorMessage = `You cannot delete ${name} because it is being used in ${productsUsingCap.length} cap(s).`;
+            const errors: ErrorInfo[] = [{ id: 'cap-card-0', errorMessage }];
+            updateSessionAttribute(req, VIEW_CAP, errors);
+            redirectTo(res, `/viewCaps?cannotDelete=${name}`);
+            return;
+        }
 
         await deleteCap(id, noc);
